# Optimize Span<T>.Fill implementation