--- a/src/libraries/System.Private.CoreLib/src/System/SpanHelpers.T.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/SpanHelpers.T.cs
@@ -4,103 +4,119 @@
 using System.Diagnostics;
 using System.Numerics;
 using System.Runtime.CompilerServices;
+using System.Runtime.Intrinsics;
 using Internal.Runtime.CompilerServices;
 
 namespace System
 {
     internal static partial class SpanHelpers // .T
     {
-        public static void Fill<T>(ref T refData, uint numElements, T value)
+        public static void Fill<T>(ref T refData, nuint numElements, T value)
         {
-            // n.b. If Fill is ever adapted to work for lengths beyond uint.MaxValue, double-check
-            // where it's used in the arithmetic operations below to ensure no integer overflow is possible.
+            // Early checks to see if it's even possible to vectorize - JIT will turn these checks into consts.
+            // - T cannot contain references (GC can't track references in vectors)
+            // - Vectorization must be hardware-accelerated
+            // - T's size must not exceed the vector's size and must be a whole power of 2
 
-            if (numElements == 0)
+            if (RuntimeHelpers.IsReferenceOrContainsReferences<T>()) { goto CannotVectorize; }
+            if (!Vector.IsHardwareAccelerated) { goto CannotVectorize; }
+            if (Unsafe.SizeOf<T>() > Vector<byte>.Count) { goto CannotVectorize; }
+            if ((Unsafe.SizeOf<T>() & (Unsafe.SizeOf<T>() - 1)) != 0) { goto CannotVectorize; } // power of 2 check
+
+            if (numElements > (uint)(Vector<byte>.Count / Unsafe.SizeOf<T>()))
             {
-                return; // nothing to do
-            }
+                // We have enough data for at least one vectorized write.
 
-            if (typeof(T) != typeof(float) && typeof(T) != typeof(double) && !RuntimeHelpers.IsBitwiseEquatable<T>())
-            {
-                goto CannotVectorize; // it might not be valid to SIMD-spray this value into the backing span
-            }
-
-            if (Vector.IsHardwareAccelerated)
-            {
-                if (numElements < (uint)(Vector<byte>.Count / Unsafe.SizeOf<T>()))
-                {
-                    goto CannotVectorize; // not enough data for even a single run through the vectorized loop
-                }
-
+                T tmp = value; // Avoid taking address of the "value" argument. It would regress performance of the loops below.
                 Vector<byte> vector;
 
-                if (typeof(T) == typeof(float))
+                if (Unsafe.SizeOf<T>() == 1)
                 {
-                    vector = (Vector<byte>)(new Vector<float>((float)(object)value!));
+                    vector = new Vector<byte>(Unsafe.As<T, byte>(ref tmp));
                 }
-                else if (typeof(T) == typeof(double))
+                else if (Unsafe.SizeOf<T>() == 2)
                 {
-                    vector = (Vector<byte>)(new Vector<double>((double)(object)value!));
+                    vector = (Vector<byte>)(new Vector<ushort>(Unsafe.As<T, ushort>(ref tmp)));
                 }
-                else
+                else if (Unsafe.SizeOf<T>() == 4)
                 {
-                    T tmp = value; // Avoid taking address of the "value" argument. It would regress performance of the loops below.
-                    if (Unsafe.SizeOf<T>() == 1)
+                    // special-case float since it's already passed in a SIMD reg
+                    vector = (typeof(T) == typeof(float))
+                        ? (Vector<byte>)(new Vector<float>((float)(object)tmp!))
+                        : (Vector<byte>)(new Vector<uint>(Unsafe.As<T, uint>(ref tmp)));
+                }
+                else if (Unsafe.SizeOf<T>() == 8)
+                {
+                    // special-case double since it's already passed in a SIMD reg
+                    vector = (typeof(T) == typeof(double))
+                        ? (Vector<byte>)(new Vector<double>((double)(object)tmp!))
+                        : (Vector<byte>)(new Vector<ulong>(Unsafe.As<T, ulong>(ref tmp)));
+                }
+                else if (Unsafe.SizeOf<T>() == 16)
+                {
+                    Vector128<byte> vec128 = Unsafe.As<T, Vector128<byte>>(ref tmp);
+                    if (Vector<byte>.Count == 16)
                     {
-                        vector = new Vector<byte>(Unsafe.As<T, byte>(ref tmp));
+                        vector = vec128.AsVector();
                     }
-                    else if (Unsafe.SizeOf<T>() == 2)
+                    else if (Vector<byte>.Count == 32)
                     {
-                        vector = (Vector<byte>)(new Vector<ushort>(Unsafe.As<T, ushort>(ref tmp)));
-                    }
-                    else if (Unsafe.SizeOf<T>() == 4)
-                    {
-                        vector = (Vector<byte>)(new Vector<uint>(Unsafe.As<T, uint>(ref tmp)));
-                    }
-                    else if (Unsafe.SizeOf<T>() == 8)
-                    {
-                        vector = (Vector<byte>)(new Vector<ulong>(Unsafe.As<T, ulong>(ref tmp)));
+                        vector = Vector256.Create(vec128, vec128).AsVector();
                     }
                     else
                     {
-                        Debug.Fail("This should never happen. Did we miss special-casing a bitwise equatable type?");
+                        Debug.Fail("Vector<T> isn't 128 or 256 bits in size?");
                         goto CannotVectorize;
                     }
                 }
+                else if (Unsafe.SizeOf<T>() == 32)
+                {
+                    vector = Unsafe.As<T, Vector256<byte>>(ref tmp).AsVector();
+                }
+                else
+                {
+                    Debug.Fail("Vector<T> is greater than 256 bits in size?");
+                    goto CannotVectorize;
+                }
 
+                ref byte refDataAsBytes = ref Unsafe.As<T, byte>(ref refData);
                 nuint totalByteLength = numElements * (nuint)Unsafe.SizeOf<T>(); // get this calculation ready ahead of time
-                nuint stopLoopAtOffset = totalByteLength & (nuint)(-Vector<byte>.Count * 2);
+                nuint stopLoopAtOffset = totalByteLength & (nuint)(nint)(2 * (int)-Vector<byte>.Count); // intentional sign extension carries the negative bit
                 nuint offset = 0;
 
                 // Loop, writing 2 vectors at a time.
+                // Compare 'numElements' rather than 'stopLoopAtOffset' because we don't want a dependency
+                // on the very recently calculated 'stopLoopAtOffset' value.
 
-                if (numElements >= 2 * (uint)(Vector<byte>.Count / Unsafe.SizeOf<T>()))
+                if (numElements >= (uint)(2 * Vector<byte>.Count / Unsafe.SizeOf<T>()))
                 {
                     do
                     {
-                        Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref Unsafe.As<T, byte>(ref refData), offset), vector);
-                        Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref Unsafe.As<T, byte>(ref refData), offset + (uint)Vector<byte>.Count), vector);
-                        offset += 2 * (uint)Vector<byte>.Count;
+                        Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref refDataAsBytes, offset), vector);
+                        Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref refDataAsBytes, offset + (nuint)Vector<byte>.Count), vector);
+                        offset += (uint)(2 * Vector<byte>.Count);
                     } while (offset < stopLoopAtOffset);
                 }
 
-                // There are [ 0, 2 * sizeof(Vector) ) bytes remaining.
-                // If there are >= sizeof(Vector) bytes remaining, write one vector now.
+                // At this point, if any data remains to be written, it's strictly less than
+                // 2 * sizeof(Vector) bytes. The loop above had us write an even number of vectors.
+                // If the total byte length instead involves us writing an odd number of vectors, write
+                // one additional vector now. The bit check below tells us if we're in an "odd vector
+                // count" situation.
 
                 if ((totalByteLength & (nuint)Vector<byte>.Count) != 0)
                 {
-                    Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref Unsafe.As<T, byte>(ref refData), offset), vector);
+                    Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref refDataAsBytes, offset), vector);
                 }
 
-                // If there's any remaining space that won't fill a full vector, write a vector at
-                // the very end of the destination buffer. This will result in overwriting a previous
-                // entry, but that's ok since we're splatting the same value for all entries, so this
-                // won't result in data corruption.
+                // It's possible that some small buffer remains to be populated - something that won't
+                // fit an entire vector's worth of data. Instead of falling back to a loop, we'll write
+                // a vector at the very end of the buffer. This may involve overwriting previously
+                // populated data, which is fine since we're splatting the same value for all entries.
+                // There's no need to perform a length check here because we already performed this
+                // check before entering the vectorized code path.
 
-                if ((totalByteLength & (nuint)(Vector<byte>.Count - 1)) != 0)
-                {
-                    Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref Unsafe.As<T, byte>(ref refData), totalByteLength - (uint)Vector<byte>.Count), vector);
-                }
+                Unsafe.WriteUnaligned(ref Unsafe.AddByteOffset(ref refDataAsBytes, totalByteLength - (nuint)Vector<byte>.Count), vector);
 
                 // And we're done!
 
@@ -109,13 +125,17 @@ namespace System
 
         CannotVectorize:
 
+            // If we reached this point, we cannot vectorize this T, or there are too few
+            // elements for us to vectorize. Fall back to an unrolled loop.
+
+            nuint i = 0;
+
+            // Write 8 elements at a time
+
+            if (numElements >= 8)
             {
-                nuint i = 0;
-                nuint stopLoopAtOffset = numElements & ~(uint)7;
-
-                // Write 8 elements at a time
-
-                for (; i < stopLoopAtOffset; i += 8)
+                nuint stopLoopAtOffset = numElements & ~(nuint)7;
+                do
                 {
                     Unsafe.Add(ref refData, (nint)i + 0) = value;
                     Unsafe.Add(ref refData, (nint)i + 1) = value;
@@ -125,34 +145,34 @@ namespace System
                     Unsafe.Add(ref refData, (nint)i + 5) = value;
                     Unsafe.Add(ref refData, (nint)i + 6) = value;
                     Unsafe.Add(ref refData, (nint)i + 7) = value;
-                }
+                } while ((i += 8) < stopLoopAtOffset);
+            }
 
-                // Write next 4 elements if needed
+            // Write next 4 elements if needed
 
-                if ((numElements & 4) != 0)
-                {
-                    Unsafe.Add(ref refData, (nint)i + 0) = value;
-                    Unsafe.Add(ref refData, (nint)i + 1) = value;
-                    Unsafe.Add(ref refData, (nint)i + 2) = value;
-                    Unsafe.Add(ref refData, (nint)i + 3) = value;
-                    i += 4;
-                }
+            if ((numElements & 4) != 0)
+            {
+                Unsafe.Add(ref refData, (nint)i + 0) = value;
+                Unsafe.Add(ref refData, (nint)i + 1) = value;
+                Unsafe.Add(ref refData, (nint)i + 2) = value;
+                Unsafe.Add(ref refData, (nint)i + 3) = value;
+                i += 4;
+            }
 
-                // Write next 2 elements if needed
+            // Write next 2 elements if needed
 
-                if ((numElements & 2) != 0)
-                {
-                    Unsafe.Add(ref refData, (nint)i + 0) = value;
-                    Unsafe.Add(ref refData, (nint)i + 1) = value;
-                    i += 2;
-                }
+            if ((numElements & 2) != 0)
+            {
+                Unsafe.Add(ref refData, (nint)i + 0) = value;
+                Unsafe.Add(ref refData, (nint)i + 1) = value;
+                i += 2;
+            }
 
-                // Write final element if needed
+            // Write final element if needed
 
-                if ((numElements & 1) != 0)
-                {
-                    Unsafe.Add(ref refData, (nint)i) = value;
-                }
+            if ((numElements & 1) != 0)
+            {
+                Unsafe.Add(ref refData, (nint)i) = value;
             }
         }