This optimizes `Span<T>.Fill` via three primary mechanisms:

- For _T = byte_, forwards directly to the _initblk_ (_memset_) implementation
- For _T = \<primitive\>_, uses a SIMD-optimized worker loop if feasible
- Removes requirement for caller to stack-spill the span argument before calling main worker API

The central SIMD loop doesn't attempt to perform any type of alignment optimization. We can consider adding this in the future if benchmarking shows this to be a worthwhile addition.

I also didn't investigate any other call sites throughout the runtime + libraries to see if they should be migrated from whatever existing code they might have to this `Span<T>.Fill` implementation. That can come as a future commit to this PR or as a future PR.

<details>
<summary>Benchmark code</summary>

```cs
[GenericTypeArguments(typeof(byte))]
[GenericTypeArguments(typeof(char))]
[GenericTypeArguments(typeof(int))]
[GenericTypeArguments(typeof(long))]
[GenericTypeArguments(typeof(float))]
[GenericTypeArguments(typeof(double))]
[GenericTypeArguments(typeof(decimal))]
[GenericTypeArguments(typeof(string))]
public class SpanFillRunner<T>
{
    private T[] _arr;

    private T _value;

    [Params(0, 3, 7, 15, 16, 24, 128, 512)]
    public int Size;

    [GlobalSetup]
    public void Setup()
    {
        _arr = new T[Size];

        _value = (T)((IConvertible)42).ToType(typeof(T), null);
    }

    [Benchmark]
    public void Fill()
    {
        var arr = _arr;
        _ = arr.Length; // prove not null
        arr.AsSpan().Fill(_value);
    }
}
```
</details>

### Benchmark results:

<details>
<summary>byte</summary>

(Note: The internal _memset_ routine uses nontemporal stores, which could explain the blazing fast runtime.)

| Method |        Job | Toolchain | Size |     Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |---------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** | **1.882 ns** | **0.0093 ns** | **0.0083 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 | 1.382 ns | 0.0320 ns | 0.0300 ns |  0.74 |    0.01 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** | **3.449 ns** | **0.0172 ns** | **0.0144 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 | 1.865 ns | 0.0445 ns | 0.0395 ns |  0.54 |    0.01 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** | **3.439 ns** | **0.0112 ns** | **0.0100 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 | 1.854 ns | 0.0195 ns | 0.0163 ns |  0.54 |    0.01 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **15** | **3.386 ns** | **0.0115 ns** | **0.0102 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   15 | 1.857 ns | 0.0219 ns | 0.0183 ns |  0.55 |    0.01 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **16** | **3.424 ns** | **0.0115 ns** | **0.0096 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   16 | 1.845 ns | 0.0181 ns | 0.0151 ns |  0.54 |    0.01 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **24** | **3.211 ns** | **0.0173 ns** | **0.0153 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   24 | 1.918 ns | 0.0093 ns | 0.0087 ns |  0.60 |    0.00 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **128** | **5.484 ns** | **0.0543 ns** | **0.0454 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  128 | 4.321 ns | 0.1117 ns | 0.1602 ns |  0.79 |    0.03 |
|        |            |           |      |          |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **512** | **7.521 ns** | **0.0423 ns** | **0.0395 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  512 | 6.451 ns | 0.0442 ns | 0.0345 ns |  0.86 |    0.01 |
</details>

<details>
<summary>char</summary>

| Method |        Job | Toolchain | Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |-----------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** |   **1.880 ns** | **0.0569 ns** | **0.0532 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 |   1.355 ns | 0.0039 ns | 0.0036 ns |  0.72 |    0.02 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** |   **3.436 ns** | **0.0137 ns** | **0.0128 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 |   2.511 ns | 0.0167 ns | 0.0131 ns |  0.73 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** |   **4.199 ns** | **0.0524 ns** | **0.0490 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 |   2.300 ns | 0.0121 ns | 0.0113 ns |  0.55 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **15** |   **5.261 ns** | **0.0325 ns** | **0.0304 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   15 |   2.591 ns | 0.0353 ns | 0.0330 ns |  0.49 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **16** |   **5.263 ns** | **0.0192 ns** | **0.0170 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   16 |   2.516 ns | 0.0150 ns | 0.0140 ns |  0.48 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **24** |   **6.809 ns** | **0.0126 ns** | **0.0106 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   24 |   2.296 ns | 0.0105 ns | 0.0098 ns |  0.34 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **128** |  **30.955 ns** | **0.1179 ns** | **0.1103 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  128 |   3.300 ns | 0.0942 ns | 0.1723 ns |  0.10 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **512** | **130.535 ns** | **1.7414 ns** | **1.5437 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  512 |   7.081 ns | 0.1504 ns | 0.1407 ns |  0.05 |    0.00 |
</details>

<details>
<summary>int</summary>

| Method |        Job | Toolchain | Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |-----------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** |   **1.603 ns** | **0.0630 ns** | **0.1183 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 |   1.720 ns | 0.1739 ns | 0.5127 ns |  0.98 |    0.23 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** |   **2.611 ns** | **0.0225 ns** | **0.0211 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 |   2.030 ns | 0.0258 ns | 0.0241 ns |  0.78 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** |   **2.638 ns** | **0.0194 ns** | **0.0181 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 |   2.062 ns | 0.0059 ns | 0.0055 ns |  0.78 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **15** |   **4.415 ns** | **0.0084 ns** | **0.0075 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   15 |   1.614 ns | 0.0045 ns | 0.0042 ns |  0.37 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **16** |   **5.104 ns** | **0.0169 ns** | **0.0149 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   16 |   2.060 ns | 0.0046 ns | 0.0041 ns |  0.40 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **24** |   **7.028 ns** | **0.0173 ns** | **0.0162 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   24 |   1.815 ns | 0.0157 ns | 0.0147 ns |  0.26 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **128** |  **29.413 ns** | **0.0861 ns** | **0.0805 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  128 |   5.971 ns | 0.0835 ns | 0.1609 ns |  0.21 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **512** | **117.060 ns** | **0.8512 ns** | **0.7962 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  512 |  17.274 ns | 0.0655 ns | 0.0512 ns |  0.15 |    0.00 |
</details>

<details>
<summary>long</summary>

| Method |        Job | Toolchain | Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |-----------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** |   **1.367 ns** | **0.0170 ns** | **0.0159 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 |   1.151 ns | 0.0086 ns | 0.0080 ns |  0.84 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** |   **2.940 ns** | **0.0849 ns** | **0.0977 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 |   2.055 ns | 0.0052 ns | 0.0044 ns |  0.69 |    0.02 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** |   **2.635 ns** | **0.0131 ns** | **0.0122 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 |   1.598 ns | 0.0203 ns | 0.0189 ns |  0.61 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **15** |   **4.427 ns** | **0.0123 ns** | **0.0115 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   15 |   1.662 ns | 0.0103 ns | 0.0097 ns |  0.38 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **16** |   **5.116 ns** | **0.0095 ns** | **0.0079 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   16 |   2.526 ns | 0.0072 ns | 0.0067 ns |  0.49 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **24** |   **6.957 ns** | **0.0298 ns** | **0.0279 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   24 |   2.974 ns | 0.0103 ns | 0.0096 ns |  0.43 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **128** |  **29.398 ns** | **0.0854 ns** | **0.0799 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  128 |  13.634 ns | 0.0477 ns | 0.0423 ns |  0.46 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **512** | **116.306 ns** | **0.1551 ns** | **0.1295 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  512 |  60.719 ns | 0.1973 ns | 0.1845 ns |  0.52 |    0.00 |
</details>

<details>
<summary>float</summary>

| Method |        Job | Toolchain | Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |-----------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** |   **1.367 ns** | **0.0189 ns** | **0.0177 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 |   1.141 ns | 0.0078 ns | 0.0073 ns |  0.83 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** |   **3.011 ns** | **0.0395 ns** | **0.0369 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 |   2.055 ns | 0.0073 ns | 0.0069 ns |  0.68 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** |   **2.742 ns** | **0.0432 ns** | **0.0404 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 |   2.054 ns | 0.0233 ns | 0.0218 ns |  0.75 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **15** |   **4.565 ns** | **0.1167 ns** | **0.1949 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   15 |   1.590 ns | 0.0297 ns | 0.0264 ns |  0.34 |    0.02 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **16** |   **4.693 ns** | **0.0448 ns** | **0.0419 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   16 |   2.078 ns | 0.0031 ns | 0.0027 ns |  0.44 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **24** |   **6.848 ns** | **0.0843 ns** | **0.0789 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   24 |   1.815 ns | 0.0172 ns | 0.0161 ns |  0.27 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **128** |  **29.045 ns** | **0.0866 ns** | **0.0810 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  128 |   6.326 ns | 0.0459 ns | 0.0429 ns |  0.22 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **512** | **115.432 ns** | **0.4247 ns** | **0.3973 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  512 |  32.200 ns | 0.1476 ns | 0.1381 ns |  0.28 |    0.00 |
</details>

<details>
<summary>double</summary>

| Method |        Job | Toolchain | Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |-----------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** |   **1.358 ns** | **0.0233 ns** | **0.0207 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 |   1.131 ns | 0.0163 ns | 0.0153 ns |  0.83 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** |   **2.956 ns** | **0.0491 ns** | **0.0459 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 |   2.024 ns | 0.0366 ns | 0.0342 ns |  0.68 |    0.02 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** |   **2.931 ns** | **0.0346 ns** | **0.0323 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 |   1.625 ns | 0.0335 ns | 0.0314 ns |  0.55 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **15** |   **4.455 ns** | **0.0533 ns** | **0.0499 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   15 |   1.678 ns | 0.0621 ns | 0.0829 ns |  0.36 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **16** |   **5.098 ns** | **0.0513 ns** | **0.0480 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   16 |   2.508 ns | 0.0802 ns | 0.1150 ns |  0.51 |    0.02 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |   **24** |   **6.853 ns** | **0.0835 ns** | **0.0781 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |   24 |   2.943 ns | 0.0378 ns | 0.0353 ns |  0.43 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **128** |  **28.694 ns** | **0.2794 ns** | **0.2613 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  128 |  13.549 ns | 0.0357 ns | 0.0298 ns |  0.47 |    0.00 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |  **512** | **115.105 ns** | **1.1767 ns** | **1.1007 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |  512 |  60.911 ns | 0.7672 ns | 0.7177 ns |  0.53 |    0.01 |
</details>

<details>
<summary>decimal</summary>

(This exercises non-SIMD code paths.)

| Method |        Job | Toolchain | Size |       Mean |     Error |    StdDev | Ratio | RatioSD |
|------- |----------- |---------- |----- |-----------:|----------:|----------:|------:|--------:|
|   **Fill** | **Job-NKOJQM** |      **main** |    **0** |   **1.387 ns** | **0.0146 ns** | **0.0136 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    0 |   1.350 ns | 0.0227 ns | 0.0212 ns |  0.97 |    0.02 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **3** |   **3.848 ns** | **0.0300 ns** | **0.0281 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    3 |   2.878 ns | 0.0370 ns | 0.0346 ns |  0.75 |    0.01 |
|        |            |           |      |            |           |           |       |         |
|   **Fill** | **Job-NKOJQM** |      **main** |    **7** |   **5.386 ns** | **0.0591 ns** | **0.0553 ns** |  **1.00** |    **0.00** |
|   Fill | Job-DYLJDS |  spanfill |    7 |   4.827 ns | 0.0740 ns | 0.0692 ns |  0.90 |    0.01 |
|        |            |           |      |            |           |           |       |         |
…