@SingleAccretion 2021-04-16
A note on the `memset` flavor used by the runtime: it is quite sensitive to alignment, at least for `x64`. [There is a comment](https://github.com/dotnet/runtime/blob/01b7e73cd378145264a7cb7a09365b41ed42b240/src/coreclr/vm/comutilnative.cpp#L626) about it in runtime source.