@GrabYourPitchforks 2021-04-16
@SingleAccretion good observation. I stepped through the memset implementation (used by initblk) and noticed that it performed alignment fixup before entering the main loop. So I wonder if this comment is stale?