@SingleAccretion 2021-04-16
> So I wonder if this comment is stale?

Very possible. My comment was the result of me measuring this a few months back, and seeing the expected 2x regression (Ivy Bridge). If it is no longer accurate, that's great!