@GrabYourPitchforks 2021-04-16
Notable changes in latest commit:

- The code now allows any non-ref-containing type that's a power of 2 and fits into a vector, including _Half_, _Guid_, _decimal_, etc.
- Cleaned up some of the write logic per feedback from @gfoidl
- Cleaned up unit test logic to avoid reflection
- Dropped the "zero elements?" early-exit check at the beginning of the method, as we generally shouldn't optimize for uncommon scenarios, and it'll eventually skip all the branches and hit method exit soon enough anyway
- Changed the `SpanHelpers` routine to take _nuint_ instead of _uint_ since it's a cheap enough change and it makes this code path more resilient to any future `NativeSpan` additions

No major changes to the overall design behind the logic.