@GrabYourPitchforks 2021-04-17
Notable change in latest commit:

Mono interpreter's implementation of initblk performs a null check on the incoming address before delegating to _memset_. Restoring the "don't call initblk for empty spans" logic seemed like a more surgical solution than trying to change the interpreter's initblk logic, which might have unintentional ripple effects. I'm only restoring this logic for mono, as coreclr's initblk logic handles the zero-length case just fine, and a zero-length input should be very rare.